GameConfig의 shipType 받는 부분 및 구현을 StageConfig로 이동
