+++ src/main/java/org/newdawn/spaceinvaders/configuration/StageConfig.java
@@ -1,7 +1,7 @@
 package org.newdawn.spaceinvaders.configuration;
 
 public class StageConfig {
-//    private String shipRef;
+    private String shipRef;
     private int shipPower;
     private int shipHealth;
     private int shipMagicPoint;
@@ -26,14 +26,15 @@ public class StageConfig {
     private int bossAlienMoveSpeed;
     private String bossAlienShotRef;
     private int bossAlienShotMoveSpeed;
-    private int stage=0;
+    private int stage;
+
 
     public StageConfig (int stage, String alienRef, int alienHealth, int alienPower, int alienRow,
                         int alienShotMoveSpeed, String bossAlienRef, int bossAlienHealth,
                         int bossAlienPower, int bossAlienMoveSpeed, String bossAlienShotRef,
-                        int bossAlienShotMoveSpeed){
+                        int bossAlienShotMoveSpeed, String shipType){
 
-        this.shipType = shipType;
+//        this.shipType = shipType;
         this.shipPower = 1;
         this.shipMagicPoint = 5;
         this.shipHealth = 3;
@@ -55,6 +56,25 @@ public class StageConfig {
         this.bossAlienMoveSpeed = bossAlienMoveSpeed;
         this.bossAlienShotRef = bossAlienShotRef;
         this.bossAlienShotMoveSpeed = bossAlienShotMoveSpeed;
+
+        switch (shipType) {
+            case ShipType.ATTACK_UP:
+                this.shipRef = "sprites/attackUpShip.png";
+                this.shipPower = shipPower + 2;
+                break;
+            case ShipType.DEFENCE_UP:
+                this.shipRef = "sprites/defenceUpShip.png";
+                this.shipHealth = shipHealth + 2;
+                break;
+            case ShipType.SPEED_UP:
+                this.shipRef = "sprites/speedUpShip.png";
+                this.shipMoveSpeed = shipMoveSpeed * 1.4;
+                this.shipShotMoveSpeed = shipShotMoveSpeed * 1.4;
+                break;
+            default:
+                this.shipRef = "sprites/ship.gif";
+                break;
+        }
     }
 
 
@@ -110,102 +130,40 @@ public class StageConfig {
     public int getBossAlienShotMoveSpeed() {
         return bossAlienShotMoveSpeed;
     }
-}
 
-/*
-public enum StageConfig {
-    STAGE1(1,"sprites/alien2.png",2, 1, 3, 300, "sprites/boss1.png", 50, 2, -300, "sprites/boss2.png", 500),
-    STAGE2(2, "sprites/alien3.png", 3,1,4,300,"sprites/boss2.png",50,2,-300,"sprites/alienshot.png",500),
-    STAGE3(3,"sprites/alien4.png",4,2,4,400,"sprites/boss3.png",50,2,-400,"sprites/alienshot.png",600);
-
-    // 변수선언
-    private int stage;
-    private String alienRef;
-    private int alienHealth;
-    private int alienPower;
-    private int alienRow;
-    private int alienShotMoveSpeed;
-    private String bossAlienRef;
-    private int bossAlienHealth;
-    private int bossAlienPower;
-    private int bossAlienMoveSpeed;
-    private String bossAlienShotRef;
-    private int bossAlienShotMoveSpeed;
-    private int stageCount=0; // static
-
-    // 생성자
-    StageConfig(int stage, String alienRef, int alienHealth, int alienPower, int alienRow,
-                int alienShotMoveSpeed, String bossAlienRef, int bossAlienHealth,
-                int bossAlienPower, int bossAlienMoveSpeed, String bossAlienShotRef,
-                int bossAlienShotMoveSpeed) {
-        this.stage = stage;
-        this.alienRef = alienRef;
-        this.alienHealth = alienHealth;
-        this.alienPower = alienPower;
-        this.alienRow = alienRow;
-        this.alienShotMoveSpeed = alienShotMoveSpeed;
-        this.bossAlienRef = bossAlienRef;
-        this.bossAlienHealth = bossAlienHealth;
-        this.bossAlienPower = bossAlienPower;
-        this.bossAlienMoveSpeed = bossAlienMoveSpeed;
-        this.bossAlienShotRef = bossAlienShotRef;
-        this.bossAlienShotMoveSpeed = bossAlienShotMoveSpeed;
+    public String getShipRef() {
+        return shipRef;
     }
 
-    // getter
-    public int getStage() {
-        return stage;
+    public int getShipPower() {
+        return shipPower;
     }
 
-    public String getAlienRef() {
-        return alienRef;
+    public int getShipHealth() {
+        return shipHealth;
     }
 
-    public int getAlienHealth() {
-        return alienHealth;
+    public int getShipMagicPoint() {
+        return shipMagicPoint;
     }
 
-    public int getAlienPower() {
-        return alienPower;
+    public double getShipMoveSpeed() {
+        return shipMoveSpeed;
     }
 
-    public int getAlienRow() {
-        return alienRow;
+    public String getShipShotRef() {
+        return shipShotRef;
     }
 
-    public int getAlienShotMoveSpeed() {
-        return alienShotMoveSpeed;
+    public double getShipShotMoveSpeed() {
+        return shipShotMoveSpeed;
     }
 
-    public String getBossAlienRef() {
-        return bossAlienRef;
+    public String getShipFirstSkillRef() {
+        return shipFirstSkillRef;
     }
 
-    public int getBossAlienHealth() {
-        return bossAlienHealth;
-    }
-
-    public int getBossAlienPower() {
-        return bossAlienPower;
-    }
-
-    public int getBossAlienMoveSpeed() {
-        return bossAlienMoveSpeed;
-    }
-
-    public String getBossAlienShotRef() {
-        return bossAlienShotRef;
-    }
-
-    public int getBossAlienShotMoveSpeed() {
-        return bossAlienShotMoveSpeed;
-    }
-
-    public int numberOfStage(){
-        for(StageConfig type : StageConfig.values()){
-            stageCount++;
-            return stageCount;
-        }
+    public String getShipType() {
+        return shipType.toString();
     }
 }
-*/
